+++ takeout_mobile/android/app/src/main/java/com/takeoutfm/mobile/MainActivity.java
@@ -1,6 +1,7 @@
-package com.defsub.takeout;
+package com.takeoutfm.mobile;
 
 import io.flutter.embedding.android.FlutterActivity;
+import com.ryanheise.audioservice.AudioServiceActivity;
 
-public class MainActivity extends FlutterActivity {
+public class MainActivity extends AudioServiceActivity {
 }