+++ takeout_watch/android/app/src/main/kotlin/com/takeoutfm/watch/MainActivity.kt
@@ -4,8 +4,9 @@ import android.os.Bundle
 import android.view.MotionEvent
 import io.flutter.embedding.android.FlutterActivity
 import com.samsung.wearable_rotary.WearableRotaryPlugin
+import com.ryanheise.audioservice.AudioServiceActivity
 
-class MainActivity : FlutterActivity() {
+class MainActivity : AudioServiceActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         intent.putExtra("background_mode", "transparent")